Add plot target to makefile, update .gitignore
A plot comparing the performance of the curent branch against refBLAST can now be generated by
make plot

Unfortunately, the y-axis scale is currently fixed - the clean implementation would be to get the peak performance from
/proc/cpuinfo

Some level3 related files and the products of 'make plot' are added to the .gitignore file.

Given the broad use of these changes, I hope for them to be cherry-picked to all level3 implemention branches.

Co-author: Dirk Nille